Update UserService.java
In the check hierarchy method there are two cases for checking whether user can update password either through reset or change password. For changing the password user should provide old and new password for processing but in the case of reset old password was not required but the updating client/user should be authorised to reset the password.